requirements.txt
akshare==1.17.7
mootdx==0.11.7
pandas==2.3.0
tqdm==4.66.4
tushare==1.4.21
scipy==1.14.1
APScheduler==3.10.4
pytz==2023.3
Flask==2.3.3
Flask-CORS==4.0.0
Pillow==10.0.1
